fix: import meta urls for assets in 3rd party dependencies
