+++ playground/assets/dep-with-asset/index.js
@@ -1,0 +1,1 @@
+export const imageUrl = new URL('./asset.png', import.meta.url).href
