+++ playground/assets/dep-with-asset/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@vitejs/test-dep-with-asset",
+  "private": true,
+  "version": "1.0.0",
+  "type": "module",
+  "main": "index.js"
+}
