+++ playground/assets/package.json
@@ -17,5 +17,8 @@
     "dev:url-base": "vite --config ./vite.config-url-base.js dev",
     "build:url-base": "vite --config ./vite.config-url-base.js build",
     "preview:url-base": "vite --config ./vite.config-url-base.js preview"
+  },
+  "dependencies": {
+    "@vitejs/test-dep-with-asset": "file:./dep-with-asset"
   }
 }
